[enhancement] Allow for more flexibility in path definition in config.
You can now use paths such as “.” or absolute paths for the local path, and it will compare the paths returned in Node’s path.resolve instead of the raw strings. This allows for more flexibility in defining your path in config and should cut down on “gotchas.”

This should fix #19, at least it did fix issues I was having with it!